1.0? Looks like it is working

+++ generator/main.py
@@ -1,16 +1,14 @@
-# TODO: apply syntax converter tmeplate
-
-
 import os
+import re
 from os.path import isfile, join
 from collections import deque as dq
+from functools import reduce
+
 
 from db import DIRS, DISTROS, LANGS
 from get_cfg import get_cfg
 
 CFG = get_cfg()
-
-
 
 
 def make_path(directory, filename, change_extension_to=False):
@@ -28,7 +26,7 @@ def make_paths(directory, filenames, change_extension_to=False):
     return dq(map(lambda f: make_path(directory, f, change_extension_to), filenames))
 
 def try_delete_file(path: str):
-    if os.path.exists(path):
+    if os.path.exists(path) and os.path.isfile(path):
         os.remove(path)
 
 
@@ -46,6 +44,25 @@ def try_split_filename_extension(s):
 def join_filename_extension(lst):
     return '{}.{}'.format(lst[0], lst[1]) 
 
+def is_valid_fn(filename):
+    return True
+    return not (len(try_split_filename_extension(filename)) != 2 
+                or filename[0] == '.'
+                or filename[-4:] == '.swp')
+
+def apply_syntax(s):
+
+
+    def get_replace_value(match): 
+        dict_path = [CFG] + match.split('.')
+        return reduce(lambda k, v: k[v.lower()], dict_path)
+
+    matches = re.findall(re.compile(r'{{{(.*?)}}}'), s)
+    for match in matches:
+        curly = '{{{' + match + '}}}'
+        s = s.replace(curly, get_replace_value(match))
+    return s 
+
 
 def gen_file(sourcefilename):
     filename, ext = try_split_filename_extension(sourcefilename)
@@ -53,14 +70,17 @@ def gen_file(sourcefilename):
     with open(make_path(DIRS['templates'], sourcefilename), 'r') as f_template:
         with open(make_path(DIRS['templates-prefixes'], ext), 'r') as f_prefix:
             with open(make_path(DIRS['do'], filename), 'w+') as f_generated:
-                dq(map(f_generated.write, [f_prefix.read(), f_template.read()] ))
+                contents = map(apply_syntax, [f_prefix.read(), f_template.read()] )
+                dq(map(f_generated.write, contents ))
 
 
 def gen_all():
 
-    sourcefilenames = list_files(DIRS['templates'])
+
+    sourcefilenames = list(filter(is_valid_fn, list_files(DIRS['templates'])))
     dq(map(try_delete_file, make_paths(DIRS['do'], sourcefilenames, change_extension_to='')))
     dq(map(gen_file, sourcefilenames))
+
 
 def main():
     gen_all()
